requirements-api.txt
# FastAPI Backend Dependencies
# ============================
# Core framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# Database
sqlalchemy>=2.0.0
aiomysql>=0.2.0
pymysql>=1.1.0
greenlet>=3.0.0

# Authentication
pyjwt>=2.8.0
bcrypt>=4.1.0
python-multipart>=0.0.6

# Development & Testing
httpx>=0.26.0
pytest>=7.4.0
pytest-asyncio>=0.23.0
aiosqlite>=0.19.0
